further english language fixes
